+++ documentation/changelog.md
@@ -69,6 +69,7 @@ The potential sequence_options are:
 * Adding hardware file of HydraHarp 400 from Pico Quant, basing on the 3.0.0.2 version of function library and user manual.
 * reworked the QDPlotter to now contain fits and a scalable number of plots. Attention: custom notebooks might break by this change.
 * Set proper minimum wavelength value in constraints of Tektronix AWG7k series HW module
+* Added a module with function to load Qudi file in a notebook
 
 
 Config changes: